+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Add the property MaxSearchResultsCount, to override the SharePoint limit of the maximum number of objects that the LDAP server returns - https://github.com/Yvand/LDAPCP/issues/209
 * Correctly initialize LDAP-specific properties with their actual value, instead of the default value of the type - https://github.com/Yvand/LDAPCP/pull/212
 * Fix an NullReferenceException in a very rare scenario where ClaimsPrincipal.Identity is null
+* Add helper methods to get/delete a directory connection in the configuration
 
 ## LDAPCP Second Edition v17.0.20240226.2 - Published in February 26, 2024